+++ components/popover/positionCalculation.js
@@ -1,6 +1,15 @@
 const ARROW_OFFSET = 7;
 const POPUP_OFFSET = 12;
 
+const POSITION_START = 'start';
+const POSITION_CENTER = 'center';
+const POSITION_END = 'end';
+
+const DIRECTION_NORTH = 'north';
+const DIRECTION_EAST = 'east';
+const DIRECTION_SOUTH = 'south';
+const DIRECTION_WEST = 'west';
+
 const getElementPositionValues = element => {
   const { top, left, right, bottom } = element.getBoundingClientRect();
 
@@ -9,8 +18,8 @@ const getElementPositionValues = element => {
     left,
     right,
     bottom,
-    middle: top + (bottom - top) / 2,
-    center: left + (right - left) / 2,
+    centerY: top + (bottom - top) / 2,
+    centerX: left + (right - left) / 2,
   };
 };
 
@@ -64,7 +73,7 @@ const getHorizontalDirectionPositionValues = ({
 });
 
 const getPositionValues = ({ direction, position, anchorPosition, popoverDimensions, inputOffsetCorrection }) =>
-  direction === 'north' || direction === 'south'
+  direction === DIRECTION_NORTH || direction === DIRECTION_SOUTH
     ? getVerticalDirectionPositionValues({
         direction,
         position,
@@ -88,37 +97,37 @@ const getHorizontalDirectionPositionTopValue = ({
   inputOffsetCorrection,
 }) => {
   switch (position) {
-    case 'middle':
-      return anchorPosition.middle - popoverDimensions.height / 2;
-    case 'top':
+    case POSITION_CENTER:
+      return anchorPosition.centerY - popoverDimensions.height / 2;
+    case POSITION_START:
       return anchorPosition.top - POPUP_OFFSET * 0.75 - inputOffsetCorrection;
-    case 'bottom':
+    case POSITION_END:
       return anchorPosition.bottom - popoverDimensions.height + POPUP_OFFSET * 0.75 + inputOffsetCorrection;
   }
 };
 
 const getHorizontalDirectionPositionLeftValue = ({ direction, anchorPosition, popoverDimensions }) =>
-  direction === 'west'
+  direction === DIRECTION_WEST
     ? anchorPosition.left - popoverDimensions.width - POPUP_OFFSET
     : anchorPosition.right + POPUP_OFFSET;
 
 const getHorizontalDirectionArrowPositionTopValue = ({ position, popoverDimensions }) => {
   switch (position) {
-    case 'middle':
+    case POSITION_CENTER:
       return popoverDimensions.height / 2 - ARROW_OFFSET;
-    case 'top':
+    case POSITION_START:
       return ARROW_OFFSET * 2.35;
-    case 'bottom':
+    case POSITION_END:
       return popoverDimensions.height - ARROW_OFFSET * 4.5;
   }
 };
 
 const getHorizontalDirectionArrowPositionLeftValue = ({ direction, popoverDimensions }) =>
-  direction === 'west' ? popoverDimensions.width - ARROW_OFFSET : -ARROW_OFFSET;
+  direction === DIRECTION_WEST ? popoverDimensions.width - ARROW_OFFSET : -ARROW_OFFSET;
 
 // VERTICAL
 const getVerticalDirectionPositionTopValue = ({ direction, anchorPosition, popoverDimensions }) =>
-  direction === 'north'
+  direction === DIRECTION_NORTH
     ? anchorPosition.top - popoverDimensions.height - POPUP_OFFSET
     : anchorPosition.bottom + POPUP_OFFSET;
 
@@ -129,52 +138,52 @@ const getVerticalDirectionPositionLeftValue = ({
   inputOffsetCorrection,
 }) => {
   switch (position) {
-    case 'center':
-      return anchorPosition.center - popoverDimensions.width / 2;
-    case 'left':
+    case POSITION_CENTER:
+      return anchorPosition.centerX - popoverDimensions.width / 2;
+    case POSITION_START:
       return anchorPosition.left - inputOffsetCorrection;
-    case 'right':
+    case POSITION_END:
       return anchorPosition.right - popoverDimensions.width + inputOffsetCorrection;
   }
 };
 
 const getVerticalDirectionArrowPositionTopValue = ({ direction, popoverDimensions }) =>
-  direction === 'north' ? popoverDimensions.height - ARROW_OFFSET : -ARROW_OFFSET;
+  direction === DIRECTION_NORTH ? popoverDimensions.height - ARROW_OFFSET : -ARROW_OFFSET;
 
 const getVerticalDirectionArrowPositionLeftValue = ({ position, popoverDimensions }) => {
   switch (position) {
-    case 'center':
+    case POSITION_CENTER:
       return popoverDimensions.width / 2 - ARROW_OFFSET;
-    case 'left':
+    case POSITION_START:
       return 2 * POPUP_OFFSET;
-    case 'right':
+    case POSITION_END:
       return popoverDimensions.width - POPUP_OFFSET * 3;
   }
 };
 
 const isInViewport = ({ direction, anchorPosition, popoverDimensions }) => {
   switch (direction) {
-    case 'north':
+    case DIRECTION_NORTH:
       return anchorPosition.top - popoverDimensions.height - POPUP_OFFSET > 0;
-    case 'south':
+    case DIRECTION_SOUTH:
       return anchorPosition.bottom + popoverDimensions.height + POPUP_OFFSET < window.innerHeight;
-    case 'east':
+    case DIRECTION_EAST:
       return anchorPosition.right + popoverDimensions.width + POPUP_OFFSET < window.innerWidth;
-    case 'west':
+    case DIRECTION_WEST:
       return anchorPosition.left - popoverDimensions.width - POPUP_OFFSET > 0;
   }
 };
 
 const getOppositeDirection = direction => {
   switch (direction) {
-    case 'north':
-      return 'south';
-    case 'south':
-      return 'north';
-    case 'east':
-      return 'west';
-    case 'west':
-      return 'east';
+    case DIRECTION_NORTH:
+      return DIRECTION_SOUTH;
+    case DIRECTION_SOUTH:
+      return DIRECTION_NORTH;
+    case DIRECTION_EAST:
+      return DIRECTION_WEST;
+    case DIRECTION_WEST:
+      return DIRECTION_EAST;
   }
 };
 
@@ -191,66 +200,64 @@ const getDirection = ({ direction, anchorPosition, popoverDimensions }) => {
     : direction;
 };
 
-const isPositionPossible = (position, anchorPosition, popoverDimensions) => {
+const isPositionPossible = (direction, position, anchorPosition, popoverDimensions) =>
+  direction === DIRECTION_NORTH || direction === DIRECTION_SOUTH
+    ? isVerticalDirectionPositionPossible(position, anchorPosition, popoverDimensions)
+    : isHorizontalDirectionPositionPossible(position, anchorPosition, popoverDimensions);
+
+const isVerticalDirectionPositionPossible = (position, anchorPosition, popoverDimensions) => {
   switch (position) {
-    case 'middle':
+    case POSITION_CENTER:
       return (
-        anchorPosition.middle + popoverDimensions.height / 2 < window.innerHeight &&
-        anchorPosition.middle - popoverDimensions.height / 2 > 0
+        anchorPosition.centerY + popoverDimensions.height / 2 < window.innerHeight &&
+        anchorPosition.centerY - popoverDimensions.height / 2 > 0
       );
-    case 'top':
+    case POSITION_START:
       return anchorPosition.top + popoverDimensions.height < window.innerHeight;
-    case 'bottom':
+    case POSITION_END:
       return anchorPosition.bottom - popoverDimensions.height > 0;
-    case 'center':
+  }
+};
+
+const isHorizontalDirectionPositionPossible = (position, anchorPosition, popoverDimensions) => {
+  switch (position) {
+    case POSITION_CENTER:
       return (
-        anchorPosition.center + popoverDimensions.width / 2 < window.innerWidth &&
-        anchorPosition.center - popoverDimensions.width / 2 > 0
+        anchorPosition.centerX + popoverDimensions.width / 2 < window.innerWidth &&
+        anchorPosition.centerX - popoverDimensions.width / 2 > 0
       );
-    case 'left':
+    case POSITION_START:
       return anchorPosition.left + popoverDimensions.width < window.innerWidth;
-    case 'right':
+    case POSITION_END:
       return anchorPosition.right - popoverDimensions.width > 0;
   }
 };
 
 const getOppositePosition = direction => {
   switch (direction) {
-    case 'top':
-      return 'bottom';
-    case 'bottom':
-      return 'top';
-    case 'left':
-      return 'right';
-    case 'right':
-      return 'left';
+    case POSITION_START:
+      return POSITION_END;
+    case POSITION_END:
+      return POSITION_START;
   }
 };
 
-const getPosition = ({ position, anchorPosition, popoverDimensions }) => {
+const getPosition = ({ direction, position, anchorPosition, popoverDimensions }) => {
   if (isPositionPossible(position, anchorPosition, popoverDimensions)) {
     return position;
   }
 
   switch (position) {
-    case 'middle':
-      return isPositionPossible('top', anchorPosition, popoverDimensions)
-        ? 'top'
-        : isPositionPossible('bottom', anchorPosition, popoverDimensions)
-          ? 'bottom'
+    case POSITION_CENTER:
+      return isPositionPossible(POSITION_START, anchorPosition, popoverDimensions)
+        ? POSITION_START
+        : isPositionPossible(POSITION_END, anchorPosition, popoverDimensions)
+          ? POSITION_END
           : position;
-    case 'center':
-      return isPositionPossible('left', anchorPosition, popoverDimensions)
-        ? 'left'
-        : isPositionPossible('right', anchorPosition, popoverDimensions)
-          ? 'right'
-          : position;
-    case 'top':
-    case 'bottom':
-      return isPositionPossible('middle', anchorPosition, popoverDimensions) ? 'middle' : getOppositePosition(position);
-    case 'left':
-    case 'right':
-      return isPositionPossible('center', anchorPosition, popoverDimensions) ? 'center' : getOppositePosition(position);
+    default:
+      return isPositionPossible(POSITION_CENTER, anchorPosition, popoverDimensions)
+        ? POSITION_CENTER
+        : getOppositePosition(position);
   }
 };
 
@@ -278,9 +285,9 @@ const getMaxPopoverHeight = ({ direction, anchorPosition, popoverContentEl }) =>
 
 const getMaxPopoverHeightValue = ({ direction, anchorPosition }) => {
   switch (direction) {
-    case 'north':
+    case DIRECTION_NORTH:
       return anchorPosition.top - POPUP_OFFSET * 2;
-    case 'south':
+    case DIRECTION_SOUTH:
       return window.innerHeight - anchorPosition.bottom - POPUP_OFFSET * 2;
   }
 };
@@ -297,7 +304,7 @@ export const calculatePositions = (
   const popoverDimensions = getElementDimensionValues(popoverEl);
 
   const direction = getDirection({ direction: inputDirection, anchorPosition, popoverDimensions });
-  const position = getPosition({ position: inputPosition, anchorPosition, popoverDimensions });
+  const position = getPosition({ direction, position: inputPosition, anchorPosition, popoverDimensions });
 
   return {
     maxPopoverHeight: getMaxPopoverHeight({