+++ stories/popover.js
@@ -1,18 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { Store, State } from '@sambego/storybook-state';
-import {
-  Banner,
-  Box,
-  Button,
-  ButtonGroup,
-  Heading3,
-  Link,
-  PopoverHorizontal,
-  PopoverVertical,
-  TextBody,
-  TextSmall,
-} from '../components';
+import { Banner, Box, Button, ButtonGroup, Heading3, Link, Popover, TextBody, TextSmall } from '../components';
 import { select, boolean, number } from '@storybook/addon-knobs/react';
 
 const store = new Store({
@@ -22,10 +11,8 @@ const store = new Store({
 const colors = ['neutral', 'mint', 'violet', 'ruby', 'gold', 'aqua', 'teal'];
 const tints = ['lightest', 'light', 'normal', 'dark', 'darkest'];
 const backdrops = ['transparent', 'dark'];
-const horizontalDirections = ['east', 'west'];
-const verticalDirections = ['north', 'south'];
-const horizontalPositions = ['top', 'middle', 'bottom'];
-const verticalPositions = ['left', 'center', 'right'];
+const directions = ['north', 'south', 'east', 'west'];
+const positions = ['start', 'center', 'end'];
 
 const handleButtonClick = event => {
   store.set({ anchorEl: event.currentTarget, active: true });
@@ -44,48 +31,6 @@ const contentBoxWithSingleTextLine = (
       </Link>{' '}
       inside
     </TextBody>
-    <TextBody>
-      This is the popover content with a{' '}
-      <Link href="#" inherit={false}>
-        link
-      </Link>{' '}
-      inside
-    </TextBody>
-    <TextBody>
-      This is the popover content with a{' '}
-      <Link href="#" inherit={false}>
-        link
-      </Link>{' '}
-      inside
-    </TextBody>
-    <TextBody>
-      This is the popover content with a{' '}
-      <Link href="#" inherit={false}>
-        link
-      </Link>{' '}
-      inside
-    </TextBody>
-    <TextBody>
-      This is the popover content with a{' '}
-      <Link href="#" inherit={false}>
-        link
-      </Link>{' '}
-      inside
-    </TextBody>
-    <TextBody>
-      This is the popover content with a{' '}
-      <Link href="#" inherit={false}>
-        link
-      </Link>{' '}
-      inside
-    </TextBody>
-    <TextBody>
-      This is the popover content with a{' '}
-      <Link href="#" inherit={false}>
-        link
-      </Link>{' '}
-      inside
-    </TextBody>
   </Box>
 );
 
@@ -95,16 +40,16 @@ storiesOf('Popover', module)
       propTablesExclude: [Link, TextBody, TextSmall, Button, State, Banner, Heading3, ButtonGroup, Box],
     },
   })
-  .add('horizontal', () => (
-    <Box>
-      <Button onClick={handleButtonClick} label="Open a horizontalPopover" />
+  .add('Basic', () => (
+    <Box display="flex" justifyContent="center">
+      <Button onClick={handleButtonClick} label="Open a basic Popover" />
       <State store={store}>
-        <PopoverHorizontal
+        <Popover
           active={false}
           backdrop={select('Backdrop', backdrops, 'transparent')}
           color={select('Color', colors, 'neutral')}
-          direction={select('Direction', horizontalDirections, 'west')}
-          position={select('Position', horizontalPositions, 'middle')}
+          direction={select('Direction', directions, 'south')}
+          position={select('Position', positions, 'center')}
           onEscKeyDown={handleCloseClick}
           onOverlayClick={handleCloseClick}
           tint={select('Tint', tints, 'lightest')}
@@ -112,41 +57,20 @@ storiesOf('Popover', module)
           offsetCorrection={number('Offset correction', 0)}
         >
           {contentBoxWithSingleTextLine}
-        </PopoverHorizontal>
-      </State>
-    </Box>
-  ))
-  .add('vertical', () => (
-    <Box>
-      <Button onClick={handleButtonClick} label="Open a vertical Popover" />
-      <State store={store}>
-        <PopoverVertical
-          active={false}
-          backdrop={select('Backdrop', backdrops, 'transparent')}
-          color={select('Color', colors, 'neutral')}
-          direction={select('Direction', verticalDirections, 'south')}
-          position={select('Position', verticalPositions, 'center')}
-          onEscKeyDown={handleCloseClick}
-          onOverlayClick={handleCloseClick}
-          tint={select('Tint', tints, 'lightest')}
-          lockScroll={boolean('Lock scroll', true)}
-          offsetCorrection={number('Offset correction', 0)}
-        >
-          {contentBoxWithSingleTextLine}
-        </PopoverVertical>
+        </Popover>
       </State>
     </Box>
   ))
   .add('with title', () => (
-    <Box>
+    <Box display="flex" justifyContent="center">
       <Button onClick={handleButtonClick} label="Open titled Popover" />
       <State store={store}>
-        <PopoverVertical
+        <Popover
           active={false}
           backdrop={select('Backdrop', backdrops, 'transparent')}
           color={select('Color', colors, 'neutral')}
-          direction={select('Direction', verticalDirections, 'south')}
-          position={select('Position', verticalPositions, 'center')}
+          direction={select('Direction', directions, 'south')}
+          position={select('Position', positions, 'center')}
           onEscKeyDown={handleCloseClick}
           onOverlayClick={handleCloseClick}
           tint={select('Tint', tints, 'lightest')}
@@ -157,20 +81,20 @@ storiesOf('Popover', module)
             <Heading3>Popover Title</Heading3>
           </Banner>
           {contentBoxWithSingleTextLine}
-        </PopoverVertical>
+        </Popover>
       </State>
     </Box>
   ))
   .add('with title & subtitle', () => (
-    <Box>
+    <Box display="flex" justifyContent="center">
       <Button onClick={handleButtonClick} label="Open titled & subtitled Popover" />
       <State store={store}>
-        <PopoverVertical
+        <Popover
           active={false}
           backdrop={select('Backdrop', backdrops, 'transparent')}
           color={select('Color', colors, 'neutral')}
-          direction={select('Direction', verticalDirections, 'south')}
-          position={select('Position', verticalPositions, 'center')}
+          direction={select('Direction', directions, 'south')}
+          position={select('Position', positions, 'center')}
           onEscKeyDown={handleCloseClick}
           onOverlayClick={handleCloseClick}
           tint={select('Tint', tints, 'lightest')}
@@ -182,20 +106,20 @@ storiesOf('Popover', module)
             <TextSmall marginTop={1}>This is the popover content</TextSmall>
           </Banner>
           {contentBoxWithSingleTextLine}
-        </PopoverVertical>
+        </Popover>
       </State>
     </Box>
   ))
   .add('with close button', () => (
-    <Box>
+    <Box display="flex" justifyContent="center">
       <Button onClick={handleButtonClick} label="Open Popover with close button" />
       <State store={store}>
-        <PopoverVertical
+        <Popover
           active={false}
           backdrop={select('Backdrop', backdrops, 'transparent')}
           color={select('Color', colors, 'neutral')}
           direction="south"
-          position={select('Position', verticalPositions, 'center')}
+          position={select('Position', positions, 'center')}
           onEscKeyDown={handleCloseClick}
           onOverlayClick={handleCloseClick}
           tint={select('Tint', tints, 'lightest')}
@@ -206,20 +130,20 @@ storiesOf('Popover', module)
             <Heading3>I am a heading 3</Heading3>
           </Banner>
           {contentBoxWithSingleTextLine}
-        </PopoverVertical>
+        </Popover>
       </State>
     </Box>
   ))
   .add('with actions', () => (
-    <Box>
+    <Box display="flex" justifyContent="center">
       <Button onClick={handleButtonClick} label="Open Popover with actions" />
       <State store={store}>
-        <PopoverVertical
+        <Popover
           active={false}
           backdrop={select('Backdrop', backdrops, 'transparent')}
           color={select('Color', colors, 'neutral')}
-          direction={select('Direction', verticalDirections, 'south')}
-          position={select('Position', verticalPositions, 'center')}
+          direction={select('Direction', directions, 'south')}
+          position={select('Position', positions, 'center')}
           onEscKeyDown={handleCloseClick}
           onOverlayClick={handleCloseClick}
           tint={select('Tint', tints, 'lightest')}
@@ -231,20 +155,20 @@ storiesOf('Popover', module)
             <Button label="Cancel" />
             <Button level="primary" label="Confirm" />
           </ButtonGroup>
-        </PopoverVertical>
+        </Popover>
       </State>
     </Box>
   ))
   .add('experiment 1', () => (
-    <Box>
+    <Box display="flex" justifyContent="center">
       <Button onClick={handleButtonClick} label="Open a experimental Popover" />
       <State store={store}>
-        <PopoverVertical
+        <Popover
           active={false}
           backdrop={select('Backdrop', backdrops, 'transparent')}
           color={select('Color', colors, 'neutral')}
-          direction={select('Direction', verticalDirections, 'south')}
-          position={select('Position', verticalPositions, 'center')}
+          direction={select('Direction', directions, 'south')}
+          position={select('Position', positions, 'center')}
           onEscKeyDown={handleCloseClick}
           onOverlayClick={handleCloseClick}
           tint={select('Tint', tints, 'lightest')}
@@ -258,20 +182,20 @@ storiesOf('Popover', module)
               et dolore magna aliquyam erat, sed diam voluptua.
             </TextBody>
           </Box>
-        </PopoverVertical>
+        </Popover>
       </State>
     </Box>
   ))
   .add('experiment 2', () => (
-    <Box>
+    <Box display="flex" justifyContent="center">
       <Button onClick={handleButtonClick} label="Open a experimental Popover" />
       <State store={store}>
-        <PopoverVertical
+        <Popover
           active={false}
           backdrop={select('Backdrop', backdrops, 'transparent')}
           color={select('Color', colors, 'neutral')}
-          direction={select('Direction', verticalDirections, 'south')}
-          position={select('Position', verticalPositions, 'center')}
+          direction={select('Direction', directions, 'south')}
+          position={select('Position', positions, 'center')}
           onEscKeyDown={handleCloseClick}
           onOverlayClick={handleCloseClick}
           tint={select('Tint', tints, 'lightest')}
@@ -284,7 +208,7 @@ storiesOf('Popover', module)
               <li>dolor sit amet</li>
             </ul>
           </Box>
-        </PopoverVertical>
+        </Popover>
       </State>
     </Box>
   ));